Add files via upload
Automation testing of Snapdeal website using Selenium and TestNG framework. To automate the testing of Snapdeal website to ensure that it is functioning correctly and meets the requirements of users.
The project uses Selenium WebDriver to automate the interaction with the Snapdeal website. TestNG framework is used to organize the tests and generate reports.